+++ package.json
@@ -1,20 +1,24 @@
 {
   "name": "fusion-plugin-react-router",
-  "description":
-    "The `fusion-plugin-react-router` package provides a universal router plugin for React.",
+  "description": "The `fusion-plugin-react-router` package provides a universal router plugin for React.",
   "version": "0.3.0",
   "license": "MIT",
   "repository": "fusionjs/fusion-plugin-react-router",
-  "files": ["dist"],
-  "main": "./dist/node.cjs.js",
-  "module": "./dist/node.es.js",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "main": "./dist/index.js",
+  "module": "./dist/index.es.js",
   "browser": {
-    "./dist/node.cjs.js": "./dist/browser.es5.cjs.js",
-    "./dist/node.es.js": "./dist/browser.es5.es.js"
+    "./dist/node.cjs.js": "./dist/browser.cjs.js",
+    "./dist/node.es.js": "./dist/browser.es.js"
   },
   "es2015": {
-    "./dist/browser.es5.cjs.js": "./dist/browser.es2015.cjs.js",
-    "./dist/browser.es5.es.js": "./dist/browser.es2015.es.js"
+    "./dist/node.cjs.js": "./dist/node.cjs.es2015.js",
+    "./dist/node.es.js": "./dist/node.es.es2015.js",
+    "./dist/browser.cjs.js": "./dist/browser.cjs.es2015.js",
+    "./dist/browser.es.js": "./dist/browser.es.es2015.js"
   },
   "es2017": {
     "./dist/browser.es5.cjs.js": "./dist/browser.es2017.cjs.js",
@@ -35,7 +39,7 @@
   },
   "devDependencies": {
     "babel-eslint": "8.2.1",
-    "create-universal-package": "2.1.1",
+    "create-universal-package": "^3.2.4",
     "eslint": "4.15.0",
     "eslint-config-fusion": "0.2.1",
     "eslint-plugin-cup": "1.0.0",
@@ -58,8 +62,7 @@
     "clean": "rm -rf dist",
     "lint": "eslint . --ignore-path .gitignore",
     "build-test": "rm -rf dist-tests && cup build-tests",
-    "just-test":
-      "node_modules/.bin/unitest --browser=dist-tests/browser.js --node=dist-tests/node.js",
+    "just-test": "node_modules/.bin/unitest --browser=dist-tests/browser.js --node=dist-tests/node.js",
     "test": "npm run build-test && npm run just-test",
     "cover": "npm run build-test && nyc npm run just-test",
     "transpile": "npm run clean && cup build",
