--- src/__tests__/__browser__/redirect.js
@@ -1,53 +1,0 @@
-// MIT License
-
-// Copyright (c) 2017 Uber Technologies, Inc.
-
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-/* eslint-env browser */
-import test from 'tape-cup';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import {Router, Route, Redirect} from '../../browser';
-
-test('test Redirect', t => {
-  const root = document.createElement('div');
-  const Hello = () => <div>Hello</div>;
-  const Moved = () => <Redirect to="/hello">Hi</Redirect>;
-  const el = (
-    <Router>
-      <div>
-        <Route path="/" component={Moved} />
-        <Route path="/hello" component={Hello} />
-      </div>
-    </Router>
-  );
-  ReactDOM.render(el, root);
-  t.ok(/Hello/.test(root.innerHTML), `matches ${root.innerHTML}`);
-  t.equal(window.location.pathname, '/hello');
-
-  // reset the url back to "/"
-  ReactDOM.render(
-    <Router>
-      <Redirect to="/" />
-    </Router>,
-    document.createElement('div')
-  );
-  t.end();
-});
